updated_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pyhd8ed1ab_0
aiohttp=3.9.3=py310h2372a71_0
aiosignal=1.3.1=pyhd8ed1ab_0
alsa-lib=1.2.8=h166bdaf_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.3=pyhd8ed1ab_0
attr=2.5.1=h166bdaf_1
attrs=23.2.0=pyh71513ae_0
blinker=1.7.0=pyhd8ed1ab_0
box2d-py=2.3.8=pypi_0
brotli=1.1.0=hd590300_1
brotli-bin=1.1.0=hd590300_1
brotli-python=1.1.0=py310hc6cd4ac_1
bzip2=1.0.8=hd590300_5
c-ares=1.27.0=hd590300_0
ca-certificates=2024.2.2=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.3=pyhd8ed1ab_0
cairo=1.16.0=ha61ee94_1014
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=py310h2fee648_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
cloudpickle=3.0.0=pyhd8ed1ab_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=py310hd41b1e2_0
cryptography=42.0.2=py310hb8475ec_0
cuda-version=11.8=h70ddcb2_3
cudatoolkit=11.8.0=h4ba93d1_13
cudnn=8.8.0.121=hcdd5f01_4
cycler=0.12.1=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
debugpy=1.8.1=py310hc6cd4ac_0
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
fftw=3.3.10=nompi_hc118613_108
flatbuffers=22.12.06=hcb278e6_2
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.49.0=py310h2372a71_0
freetype=2.12.1=h267a509_2
frozenlist=1.4.1=py310h2372a71_0
gast=0.4.0=pyh9f0ad1d_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
glib=2.78.4=hfc55251_4
glib-tools=2.78.4=hfc55251_4
google-auth=2.28.2=pyhca7485f_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.51.1=py310h4a5735c_1
gst-plugins-base=1.22.0=h4243ec0_2
gstreamer=1.22.0=h25f0c4b_2
gstreamer-orc=0.4.38=hd590300_0
gym=0.26.1=py310hfdc917e_0
gym-notices=0.0.8=pyhd8ed1ab_0
h5py=3.11.0=pypi_0
harfbuzz=6.0.0=h8e241bc_0
hdf5=1.14.1=nompi_h4f84152_100
icu=70.1=h27087fc_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
ipykernel=6.29.3=pyhd33586a_0
ipympl=0.9.3=pyhd8ed1ab_0
ipython=8.22.1=pyh707e725_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.1.2=pyhd8ed1ab_0
jack=1.9.22=h11f4161_0
jedi=0.19.1=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py310hff52083_0
jupyterlab_widgets=3.0.10=pyhd8ed1ab_0
keras=3.3.3=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=py310hd41b1e2_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libabseil=20220623.0=cxx17_h05df665_6
libaec=1.1.2=h59595ed_1
libblas=3.9.0=21_linux64_openblas
libbrotlicommon=1.1.0=hd590300_1
libbrotlidec=1.1.0=hd590300_1
libbrotlienc=1.1.0=hd590300_1
libcap=2.67=he9d0100_0
libcblas=3.9.0=21_linux64_openblas
libclang=18.1.1=pypi_0
libclang13=15.0.7=default_ha2b6cf4_4
libcups=2.3.3=h36d4200_3
libcurl=8.1.2=h409715c_0
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libevent=2.1.10=h28343ad_4
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libflac=1.4.3=h59595ed_0
libgcc-ng=13.2.0=h807b86a_5
libgcrypt=1.10.3=hd590300_0
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libglib=2.78.4=hf2295e7_4
libgomp=13.2.0=h807b86a_5
libgpg-error=1.48=h71f35ed_0
libgrpc=1.51.1=h4fad500_1
libiconv=1.17=hd590300_2
liblapack=3.9.0=21_linux64_openblas
libllvm14=14.0.6=hcd5def8_4
libllvm15=15.0.7=hadd5161_1
libnghttp2=1.58.0=h47da74e_0
libnsl=2.0.1=hd590300_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.26=pthreads_h413a1c8_0
libopus=1.3.1=h7f98852_1
libpng=1.6.43=h2797004_0
libpq=15.3=hbcd7760_1
libprotobuf=3.21.12=hfc55251_2
libsndfile=1.2.2=hc60ed4a_1
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.45.1=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-ng=13.2.0=h7e041cc_5
libsystemd0=253=h8c4010b_1
libtiff=4.5.0=h6adf6a1_2
libtool=2.4.7=h27087fc_0
libudev1=253=h0b41bf4_0
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.13=h7f98852_1004
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.5.0=h79f4944_1
libxml2=2.10.3=hca2bb57_4
libzlib=1.2.13=hd590300_5
lz4-c=1.9.4=hcb278e6_0
markdown=3.5.2=pyhd8ed1ab_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=py310h2372a71_0
matplotlib=3.8.3=py310hff52083_0
matplotlib-base=3.8.3=py310h62c0568_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.3.2=pypi_0
mpg123=1.32.4=h59595ed_0
multidict=6.0.5=py310h2372a71_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.33=hf1915f5_6
mysql-libs=8.0.33=hca2cd23_6
namex=0.0.8=pypi_0
nccl=2.20.5.1=h6103f9b_0
ncurses=6.4=h59595ed_2
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pyhd8ed1ab_0
nspr=4.35=h27087fc_0
nss=3.98=h1d7d5a4_0
numpy=1.26.4=py310hb13e2d6_0
oauthlib=3.2.2=pyhd8ed1ab_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.1.5=hd590300_0
opt_einsum=3.3.0=pyhc1e730c_2
optree=0.11.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pcre2=10.43=hcad00b1_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310h023d228_1
pip=24.0=pyhd8ed1ab_0
pixman=0.43.2=h59595ed_0
platformdirs=4.2.0=pyhd8ed1ab_0
ply=3.11=py_1
portaudio=19.6.0=h583fa2b_7
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.21.12=py310heca2aa9_0
psutil=5.9.8=py310h2372a71_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pulseaudio=16.1=hcb278e6_3
pulseaudio-client=16.1=h5195f5e_3
pulseaudio-daemon=16.1=ha8d29e2_3
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pyhd8ed1ab_0
pyasn1-modules=0.3.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygame=2.5.2=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.8.0=pyhd8ed1ab_1
pyopenssl=24.0.0=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyqt=5.15.7=py310hab646b1_3
pyqt5-sip=12.11.0=py310heca2aa9_3
pysocks=1.7.1=pyha2e5f31_6
python=3.10.13=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-flatbuffers=24.3.6=pyh59ac667_0
python_abi=3.10=4_cp310
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=25.1.2=py310h795f18f_0
qt-main=5.15.8=h5d23da1_6
re2=2023.02.01=hcb278e6_0
readline=8.2=h8228510_1
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rich=13.7.1=pypi_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.12.0=py310hb13e2d6_2
setuptools=69.1.1=pyhd8ed1ab_0
sip=6.7.12=py310hc6cd4ac_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h9fff704_0
stack_data=0.6.2=pyhd8ed1ab_0
swig=4.2.1=pypi_0
tensorboard=2.16.2=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow-cpu=2.16.1=pypi_0
tensorflow-io-gcs-filesystem=0.37.0=pypi_0
termcolor=2.4.0=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
tornado=6.4=py310h2372a71_0
traitlets=5.14.1=pyhd8ed1ab_0
typing-extensions=4.10.0=hd8ed1ab_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024a=h0c530f3_0
unicodedata2=15.1.0=py310h2372a71_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.1=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
widgetsnbextension=4.0.10=pyhd8ed1ab_0
wrapt=1.16.0=py310h2372a71_0
xcb-util=0.4.0=h516909a_0
xcb-util-image=0.4.0=h166bdaf_0
xcb-util-keysyms=0.4.0=h516909a_0
xcb-util-renderutil=0.3.9=h166bdaf_0
xcb-util-wm=0.4.1=h516909a_0
xkeyboard-config=2.38=h0b41bf4_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.4=h0b41bf4_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
yarl=1.9.4=py310h2372a71_0
zeromq=4.3.5=h59595ed_1
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
